+++ packages/base/src/theming/getEffectiveLinksHrefs.ts
@@ -2,7 +2,7 @@ import { getUrl } from "../CSP.js";
 import { getFeature } from "../FeaturesRegistry.js";
 import type UI5Element from "../UI5Element.js";
 import type OpenUI5Enablement from "../features/OpenUI5Enablement.js";
-import { ComponentStylesData, StyleDataCSP } from "../types.js";
+import type { ComponentStylesData, StyleDataCSP } from "../types.js";
 
 const MAX_DEPTH_INHERITED_CLASSES = 10; // TypeScript complains about Infinity and big numbers