+++ packages/main/src/Input.ts
@@ -4,7 +4,7 @@ import slot from "@ui5/webcomponents-base/dist/decorators/slot.js";
 import event from "@ui5/webcomponents-base/dist/decorators/event.js";
 import customElement from "@ui5/webcomponents-base/dist/decorators/customElement.js";
 import languageAware from "@ui5/webcomponents-base/dist/decorators/languageAware.js";
-import type { ClassMap } from "@ui5/webcomponents-base/dist/types.js";
+import type { ClassMap, ComponentStylesData } from "@ui5/webcomponents-base/dist/types.js";
 import litRender from "@ui5/webcomponents-base/dist/renderer/LitRenderer.js";
 import ResizeHandler from "@ui5/webcomponents-base/dist/delegate/ResizeHandler.js";
 import type { ResizeObserverCallback } from "@ui5/webcomponents-base/dist/delegate/ResizeHandler.js";
@@ -270,7 +270,7 @@ class Input extends UI5Element implements SuggestionComponent, IFormElement {
 	 * @type {boolean}
 	 * @defaultvalue false
 	 * @private
-	 * @sicne 1.0.0-rc.8
+	 * @since 1.0.0-rc.8
 	 */
 	@property({ type: Boolean })
 	highlight!: boolean;
@@ -654,7 +654,7 @@ class Input extends UI5Element implements SuggestionComponent, IFormElement {
 		return InputPopoverTemplate;
 	}
 
-	static get styles() {
+	static get styles(): ComponentStylesData {
 		return styles;
 	}