+++ packages/main/src/TabContainer.ts
@@ -673,13 +673,13 @@ class TabContainer extends UI5Element {
 	 * The order of tabs is depth-first. For example, given the following slotted elements:
 	 * <code>
 	 * 	<ui5-tabcontainer>
-	 * 		<ui5-tab id="First">
+	 * 		<ui5-tab id="First" text="First">
 	 * 			...
-	 * 			<ui5-tab slot="subTabs" id="Nested">...</ui5-tab>
+	 * 			<ui5-tab slot="subTabs" id="Nested" text="Nested">...</ui5-tab>
 	 * 		</ui5-tab>
-	 * 		<ui5-tab id="Second">...</ui5-tab>
+	 * 		<ui5-tab id="Second" text="Second">...</ui5-tab>
 	 * 		<ui5-tab-separator id="sep"></ui5-tab-separator>
-	 * 		<ui5-tab id="Third">...</ui5-tab>
+	 * 		<ui5-tab id="Third" text="Third">...</ui5-tab>
 	 * 	</ui5-tabcontainer>
 	 * </code>
 	 * Calling <code>allItems</code> on this TabContainer will return the instances in the following order: